Move generate transaction out of node and into libs/transaction.
